mc_bot/requirements.txt
python-dotenv
python-telegram-bot
requests
Pillow